Create tests for common.Search().
Using the path_test.go file built by the Golang
devs as a base, tests have been created to verify
the functionality of common.Search().

A mock file system is created and fake test files
are generated to see if they get picked up by
common.Search().

Also included in this CL is a bug fix for
proctor-nodejs that was discovered using this test.
proctor-nodejs used to allow multiple "-" in its
test name filter. The regex has been updated to
prevent this.

PiperOrigin-RevId: 261950746